Optimized code in aprgalgorithm module

+++ AprgAlgorithm/AprgAlgorithm/tst/AlgorithmTests/String/SubstringSearch/RabinKarpSubstringSearchWithRunningHash_unit.cpp
@@ -12,14 +12,12 @@ namespace algorithm {
 
 namespace {
 using Index = int;
-using HashValue = unsigned long long;
+using HashValue = uint64_t;
 using SubstringSearchForTest = RabinKarpSubstringSearchWithRunningHash<Index, HashValue>;
 }  // namespace
-
 TEST(RabinKarpSubstringSearchWithRunningHashTest, SearchWorksUsingEmptyStringsAndDoesNotCrash) {
     testSearchUsingEmptyStringsAndDoesNotCrash<SubstringSearchForTest, Index>();
 }
-
 TEST(RabinKarpSubstringSearchWithRunningHashTest, SearchWorksUsingExample1) {
     testSearchUsingExample1<SubstringSearchForTest, Index>();
 }
