+++ AprgAlgorithm/AprgAlgorithm/tst/AlgorithmTests/String/SubstringSearch/RabinKarpSubstringSearchWithSubstringHash_unit.cpp
@@ -12,13 +12,11 @@ namespace algorithm {
 
 namespace {
 using Index = int;
-using HashValue = unsigned long long;
+using HashValue = uint64_t;
 using SubstringSearchForTest = RabinKarpSubstringSearchWithSubstringHash<Index, HashValue>;
 }  // namespace
-
 TEST(RabinKarpSubstringSearchWithSubstringHashTest, SearchWorksUsingEmptyStringsAndDoesNotCrash) {
     // SubstringSearchUsingZArray has unique results of handling empty strings
-
     string emptyString;
     string nonEmptyString("NonEmptyString");
     SubstringSearchForTest emptyStringSearch(emptyString);
